Fix channel posting feature

+++ src/cogs/cm_auto_post/ui.py
@@ -113,7 +113,7 @@ class PostChannelView(View):
             options = [discord.SelectOption(label=f"Image {idx + 1}", description=image, value=image) for idx, image in enumerate(images)]
             self.add_item(
                 Select(
-                    min_values=1,
+                    min_values=0,
                     max_values=len(options),
                     options=options,
                     placeholder="Choose image(s)",
@@ -137,7 +137,7 @@ class PostChannelView(View):
             options = cmap_conf.generate_post_channel_options()
             self.add_item(
                 Select(
-                    min_values=1,
+                    min_values=0,
                     max_values=len(options),
                     options=options,
                     placeholder="Choose post channel(s)",
@@ -149,9 +149,27 @@ class PostChannelView(View):
             )
 
         self.is_confirmed = False
+        self.images = images
 
     @discord.ui.button(label="Confirm", style=discord.ButtonStyle.green, custom_id="confirm", emoji="✔", row=3)
     async def confirm(self, interaction: discord.Interaction, *_):
+        # TODO: Add a check here to see whether user actually selected anything
+        missing_fields = []
+        if self.images is not None: 
+            if not dict_has_key(self.dict_values, "post_channel_image_select"):
+                missing_fields.append("image(s)")
+            elif len(self.dict_values["post_channel_image_select"]) == 0:
+                missing_fields.append("image(s)")
+
+        if not dict_has_key(self.dict_values, "post_channel_select"):
+            missing_fields.append("channel(s)")
+        elif len(self.dict_values["post_channel_select"]) == 0:
+            missing_fields.append("channel(s)")
+        
+        if len(missing_fields) != 0:
+            await interaction.response.send_message(content=f"Please select {' and '.join(missing_fields)} to create post", ephemeral=True)
+            return
+
         self.is_confirmed = True
         self.interaction = interaction
         self.stop()
@@ -167,11 +185,10 @@ class PostEmbed(discord.Embed):
     def __init__(self, post_details: Optional[dict] = None, *args, **kwargs):
         super().__init__(*args, **kwargs)
         
-        self.add_field(name="Event Details", value=post_details["event_details"] if post_details is not None and dict_has_key(post_details, "event_details") else "-No event details entered-", inline=False)
-        self.add_field(name="Custom Caption", value=post_details["caption"] if post_details is not None and dict_has_key(post_details, "caption") else "-No custom caption entered-", inline=False)
-
-        # TODO: Add previews row
         caption = CMAutoPostConfig.generate_post_caption(post_details)
+
+        self.add_field(name="Event Details", value=f'{post_details["event_details"]}\n\u200B' if post_details is not None and dict_has_key(post_details, "event_details") else "-No event details entered-\n\u200B", inline=False)
+        self.add_field(name="Custom Caption", value=f'{post_details["caption"]}\n\u200B' if post_details is not None and dict_has_key(post_details, "caption") else "-No custom caption entered-\n\u200B", inline=False)
         self.add_field(name="Caption Preview", value=caption if caption is not None else "-No preview generated-", inline=False)
 
 
@@ -190,7 +207,7 @@ class ClearButton(discord.ui.Button):
     async def callback(self, interaction: discord.Interaction):
         del self.view.post_details[self.custom_id.replace('_clear', '')]
 
-        await self.view.embedded_message.edit(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld", post_details=self.view.post_details))
+        await self.view.embedded_message.edit(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld\n\u200B", post_details=self.view.post_details))
 
         await interaction.response.defer()
 
@@ -222,29 +239,50 @@ class PostView(View):
                 )
             )
 
+        self.input_message = None
+        self.is_cancelled = False
+
 
     async def retrieve_user_input(self, interaction: discord.Interaction, button_id: str):
         cmap_conf = CMAutoPostConfig()
 
         feed_channel = await interaction.guild.fetch_channel(cmap_conf.data["config"]["feed_channel_id"])
-        view = CancelView(timeout=90)
-        message = await feed_channel.send(content=f"⚠️ Please enter the {self.input_names[button_id]}", view=view)  
+        
+        embed = discord.Embed(title=f"Enter the {self.input_names[button_id]}", description=f"The next message you send in <#{feed_channel.id}> will be recorded as the {self.input_names[button_id]}")
+        embed.set_author(name=interaction.user.display_name, icon_url=interaction.user.avatar)
+        embed.set_footer(text="Data is recorded successfully when the previous embed is updated with the data.")
+
+        view = CancelView(timeout=30)
+        message = await feed_channel.send(embed=embed, view=view) # TODO: Do an embed
+
+        self.input_message = message
 
         finished, unfinished = await asyncio.wait([self.bot.wait_for('message', check=lambda message: message.author == interaction.user and message.channel == feed_channel), view.wait()], return_when=asyncio.FIRST_COMPLETED)
-        
+
+        for task in unfinished:
+            task.cancel()
+
+        # TODO: Keep track of unfinished tasks and message sent, when the cancel button is clicked, delete all of that and remove the coroutines
+        if self.is_cancelled:
+            print("Cancelled effective")
+            return
+
         if isinstance(list(finished)[0].result(), discord.Message):
             user_input = list(finished)[0].result()
             self.post_details[button_id] = user_input.content
-            await message.delete()
             await user_input.delete()
-
+            await message.delete()
+            
             # TODO: Add Twitter link to the Embed
-            await self.embedded_message.edit(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld", post_details=self.post_details))
-        else:
-            await view.interaction.response.send_message(content=f"The {self.input_names[button_id]} was not recorded", ephemeral=True)
+            await self.embedded_message.edit(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld\n\u200B", post_details=self.post_details))
+            return
+
+        elif list(finished)[0].result(): # True signifies that it is timed out
+            await view.interaction.response.send_message(content=f"The user input timed out, please try again!", ephemeral=True)
+        else: # False signifies the cancel button was clicked
+            await view.interaction.response.send_message(content=f"The {self.input_names[button_id]} was not entered", ephemeral=True)
         
-        for task in unfinished:
-            task.cancel()
+        await message.delete()
         
 
     @discord.ui.button(label="Event Details", style=discord.ButtonStyle.primary, custom_id="event_details", emoji="📆", row=0)
@@ -252,98 +290,34 @@ class PostView(View):
         await interaction.response.defer()
         await self.retrieve_user_input(interaction=interaction, button_id=button.custom_id)
 
-    @discord.ui.button(label="Custom Caption", style=discord.ButtonStyle.primary, custom_id="caption", emoji="⚠️", row=3)
+    @discord.ui.button(label="Custom Caption", style=discord.ButtonStyle.primary, custom_id="caption", emoji="⚠️", row=1)
     async def custom(self, interaction: discord.Interaction, button: discord.ui.Button):
         await interaction.response.defer()
         await self.retrieve_user_input(interaction=interaction, button_id=button.custom_id)
 
-    @discord.ui.button(label="Save", style=discord.ButtonStyle.green, custom_id="save", emoji="✔", row=4)
-    async def save(self, interaction: discord.Interaction, *_):
+    @discord.ui.button(label="Post", style=discord.ButtonStyle.green, custom_id="post", emoji="📮", row=2)
+    async def post(self, interaction: discord.Interaction, *_):
+        # TODO: Check whether caption is generated
+        caption = CMAutoPostConfig.generate_post_caption(self.post_details)
+
+        if caption is None:
+            await interaction.response.send_message(content="Please enter a caption before posting", ephemeral=True)
+            return
+
         self.is_confirmed = True
         self.interaction = interaction
         self.stop()
 
-    @discord.ui.button(label="Cancel", style=discord.ButtonStyle.red, custom_id="cancel", emoji="✖️", row=4)
+    @discord.ui.button(label="Cancel", style=discord.ButtonStyle.red, custom_id="cancel", emoji="✖️", row=2)
     async def cancel(self, interaction: discord.Interaction, *_):
+        if self.input_message is not None:
+            print("Deleting message")
+            await self.input_message.delete()
+
+        self.is_cancelled = True
         self.is_confirmed = False
         self.interaction = interaction
         self.stop()
-
-
-class PersistentPostChannelButton(discord.ui.Button):
-    def __init__(
-        self,
-        *,
-        style: discord.ButtonStyle = discord.ButtonStyle.primary,
-        label: Optional[str] = None,
-        disabled: bool = False,
-        custom_id: Optional[str] = None,
-        url: Optional[str] = None,
-        emoji: Optional[Union[str, discord.Emoji, discord.PartialEmoji]] = None,
-        row: Optional[int] = None,
-        value: Any,
-        files: List[discord.File]
-    ):
-        super().__init__(
-            style=style, label=label, disabled=disabled, custom_id=custom_id, url=url, emoji=emoji, row=row
-        )
-        self.value = value
-        self.files = files
-
-    
-    async def callback(self, interaction: discord.Interaction):
-        if self.value == "stop":
-            self.view.stop()
-            self.view.interaction = interaction
-            return
-
-        post_modal = PostModal(
-            title="Create Post",
-            custom_id="create_post_modal",
-            success_msg="The post details was successfully recorded!",
-            error_msg="A few problems were encountered when recording the post details, please try again!",
-        )
-
-        await interaction.response.send_modal(post_modal)
-        timeout = await post_modal.wait()
-
-        if timeout:
-            await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
-            return
-        
-        # Post
-        post_channel = await interaction.guild.fetch_channel(int(self.value))
-        post_details = post_modal.get_values()
-        content = f'```ml\n{post_details["date"]} {post_details["name"]} | cr: {post_details["handle"]} (@INSERT_TWITTER_HANDLE)\n```'
-        await post_channel.send(content=content, files=[discord.File(fp=f"loonatheworld/{f}") for f in self.files])
-        await interaction.followup.send(content=f"Post successfully created in <#{post_channel.id}>")
-
-
-class PersistentPostView(View):
-    def __init__(self, *, timeout: Optional[float] = None, **kwargs):
-        super().__init__(timeout=timeout)
-
-        if kwargs != {}:
-            self.message_id = kwargs["message_id"]
-
-            cmap_conf = CMAutoPostConfig()
-
-            for channel in cmap_conf.post_channels:
-                self.add_item(
-                    PersistentPostChannelButton(
-                        label=channel["label"], value=channel["id"], custom_id=f"persistent:{self.message_id}:{channel['name']}", files=kwargs["files"]
-                    )
-                )
-
-            self.add_item(
-                PersistentPostChannelButton(
-                    style=discord.ButtonStyle.red,
-                    custom_id=f"persistent:{self.message_id}:stop",
-                    emoji="✖️",
-                    value="stop",
-                    files=kwargs["files"]
-                )
-            )
 
 
 class PersistentTweetButton(discord.ui.Button):
@@ -373,23 +347,6 @@ class PersistentTweetButton(discord.ui.Button):
         }
         self.bot = bot
 
-    async def send_post_modal(self, interaction: discord.Interaction):
-        post_modal = PostModal(
-            title="Create Post",
-            custom_id="create_post_modal",
-            success_msg="The post details was successfully recorded!",
-            error_msg="A few problems were encountered when recording the post details, please try again!",
-        )
-
-        await interaction.response.send_modal(post_modal)
-        timeout = await post_modal.wait()
-
-        if timeout:
-            await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
-            return
-        
-        return post_modal.get_values(), post_modal.interaction
-
     async def send_post_channel_view(self, interaction: discord.Interaction, images: Optional[List[str]] = None):
         post_channel_view = PostChannelView(
             timeout=90,
@@ -410,81 +367,44 @@ class PersistentTweetButton(discord.ui.Button):
 
         if timeout:
             await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
-            return None
-        elif not post_channel_view.is_confirmed and post_channel_view.dict_values == {}:
+        elif not post_channel_view.is_confirmed:
             await interaction.followup.send(content="No post was sent", ephemeral=True)
-            return None
-        
-        return (post_channel_view.dict_values, post_channel_view.interaction)
+        else:
+            await self.send_post_view(post_channel_view.interaction, post_channel_view.dict_values, images)
+
+
+    async def send_post_view(self, interaction: discord.Interaction, values: dict, images: Optional[List[str]] = None):
+        post_channel_ids = values["post_channel_select"]
+        images_to_post = values["post_channel_image_select"] if images is not None and self.callback_type == "select" else self.filenames
+
+        await interaction.response.send_message(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld\n\u200B"))
+        embedded_message = await interaction.original_response()
+        view = PostView(bot=self.bot, embedded_message=embedded_message, timeout=90)
+        await interaction.edit_original_response(view=view)
+
+        timeout = await view.wait()
+
+        await interaction.edit_original_response(view=None)
+
+        if timeout:
+            await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
+        elif view.is_confirmed:
+            caption = CMAutoPostConfig.generate_post_caption(view.post_details)
+
+            post_channels = []
+            for post_channel_id in post_channel_ids:
+                post_channel = await interaction.guild.fetch_channel(int(post_channel_id))
+                await post_channel.send(content=caption, files=[discord.File(fp=f"loonatheworld/{fp}") for fp in images_to_post])
+                post_channels.append(f"<#{post_channel.id}>")
+
+            await interaction.followup.send(content=f"Post(s) successfully created in {', '.join(post_channels)}")
+    
 
     async def select(self, interaction: discord.Interaction):
-        res = await self.send_post_channel_view(interaction, self.filenames)
-
-        if res is None:
-            return
-
-        post_channel, post_channel_view_interaction = res
-        post_channel_ids = post_channel["post_channel_select"]
-        selected_images = post_channel["post_channel_image_select"]
-
-        await post_channel_view_interaction.response.send_message(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld"))
-        embedded_message = await post_channel_view_interaction.original_response()
-        view = PostView(bot=self.bot, embedded_message=embedded_message, timeout=90)
-        await post_channel_view_interaction.edit_original_response(view=view)
-        
-        timeout = await view.wait()
-
-        await post_channel_view_interaction.edit_original_response(view=None)
-
-        if timeout:
-            await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
-        elif view.is_confirmed:
-            caption = CMAutoPostConfig.generate_post_caption(view.post_details)
-
-            if caption is not None:
-                channels = []
-                for post_channel_id in post_channel_ids:
-                    channel = await interaction.guild.fetch_channel(int(post_channel_id))
-                    await channel.send(content=caption, files=[discord.File(fp=f"loonatheworld/{f}") for f in selected_images])
-                    channels.append(f"<#{channel.id}>")
-
-                await post_channel_view_interaction.followup.send(content=f"Post(s) successfully created in {', '.join(channels)}")
-            else:
-                await post_channel_view_interaction.followup.send(content=f"No post(s) created in {', '.join(channels)} as no captions were provided")
-        else:
-            await interaction.followup.send(content="Posting cancelled", ephemeral=True)
-        
+        await self.send_post_channel_view(interaction, self.filenames)
 
     async def all(self, interaction: discord.Interaction):
-        post_channel, post_channel_view_interaction = await self.send_post_channel_view(interaction)
-        post_channel_ids = post_channel["post_channel_select"]
-
-        await post_channel_view_interaction.response.send_message(embed=PostEmbed(title="Post Details", description="Post details for post INSERT TWITTER LINK HERE by @loonatheworld"))
-        embedded_message = await post_channel_view_interaction.original_response()
-        view = PostView(bot=self.bot, embedded_message=embedded_message, timeout=90)
-        await post_channel_view_interaction.edit_original_response(view=view)
-        
-        timeout = await view.wait()
-
-        await post_channel_view_interaction.edit_original_response(view=None)
-
-        if timeout:
-            await interaction.followup.send(content="The command has timed out, please try again!", ephemeral=True)
-        elif view.is_confirmed:
-            caption = CMAutoPostConfig.generate_post_caption(view.post_details)
-
-            if caption is not None:
-                channels = []
-                for post_channel_id in post_channel_ids:
-                    channel = await post_channel_view_interaction.guild.fetch_channel(int(post_channel_id))
-                    await channel.send(content=caption, files=[discord.File(fp=f"loonatheworld/{f}") for f in self.filenames])
-                    channels.append(f"<#{channel.id}>")
-
-                await post_channel_view_interaction.followup.send(content=f"Post(s) successfully created in {', '.join(channels)}")
-            else:
-                await post_channel_view_interaction.followup.send(content=f"No post(s) created in {', '.join(channels)} as no captions were provided")
-        else:
-            await interaction.followup.send(content="Posting cancelled", ephemeral=True)
+        await self.send_post_channel_view(interaction)
 
     async def stop(self, interaction: discord.Interaction):
         self.view.stop()
